Fix all repeat albums being filtered out
There is a silly mistake in the repeat album filter that causes all albums to be filtered out, if repeats_prefer_explicit is set to True.

It fails to check if all albums in the group are not explicit, which is very common because most albums are not explicit.

This commit fixes that by only filtering out the non-explicit albums, if there is at least one explicit album in the group.